requirements.txt
pypdf==3.15.1
pdfplumber==0.10.2
Pillow==10.1.0
cryptography==41.0.4
pytest==7.4.3
pytest-cov==4.1.0
pyinstaller==6.1.0 